FindHotels/requirements.txt
python==3.9.1
django==3.1.4
django-import-export
